+++ ReleaseNotes.md
@@ -5,6 +5,7 @@
 - hidden field fix
 - plan HTML fix and added to [complex layout example](https://github.com/ma-ha/easy-web-app/tree/master/examples/complex-layout)
 - fix ignored page layouts in nav
+- table: fix editable checkbox 
 
 ## 0.17.x
 - Table/List: support number